fix: get route assets by route manifest
