+++ .changeset/quick-grapes-breathe.md
@@ -1,0 +1,6 @@
+---
+'@ice/route-manifest': patch
+'@ice/app': patch
+---
+
+fix: get route assets by route manifest
