+++ examples/memory-router/ice.config.mts
@@ -4,7 +4,7 @@ export default defineConfig(() => ({
   publicPath: '/',
   ssr: false,
   ssg: false,
-  splitChunks: false,
+  codeSplitting: 'page',
   routes: {
     injectInitialEntry: true,
   }
