+++ examples/memory-router/src/pages/dashboard/index.tsx
@@ -1,0 +1,7 @@
+export default function Dashboard() {
+  return (
+    <div>
+      <h2>dashboard</h2>
+    </div>
+  );
+}
