+++ packages/runtime/matchRoutes.d.ts
@@ -1,0 +1,3 @@
+import matchRoutes from './esm/matchRoutes';
+
+export default matchRoutes;
